requirements.txt
boto3
mysql-connector-python
tqdm
elasticsearch
fastapi
uvicorn
psycopg2
opensearch-py 
requests-aws4auth
nltk

# For testing
python-dotenv==1.0.1
